rkt, flag: Move optionList, bitFlags and secFlags to flag
This simply moves the code from one place to another. For that to
work, all the types and variables are now exported.

In following commits, variables and BitFlags will become private.
