+++ rkt/cli_apps_test.go
@@ -138,115 +138,3 @@ func TestParsePortFlag(t *testing.T) {
 		}
 	}
 }
-
-var options = []string{"zero", "one", "two"}
-
-func TestOptionList(t *testing.T) {
-	tests := []struct {
-		opts string
-		ex   string
-		err  bool
-	}{
-		{
-			opts: "zero,two",
-			ex:   "zero,two",
-			err:  false,
-		},
-		{ // Duplicate test
-			opts: "one,two,two",
-			ex:   "",
-			err:  true,
-		},
-		{ // Not permissible test
-			opts: "one,two,three",
-			ex:   "",
-			err:  true,
-		},
-		{ // Empty string
-			opts: "",
-			ex:   "",
-			err:  false,
-		},
-	}
-
-	for i, tt := range tests {
-		// test newOptionsList
-		if _, err := newOptionList(options, tt.opts); (err != nil) != tt.err {
-			t.Errorf("test %d: unexpected error in newOptionList: %v", i, err)
-		}
-
-		// test optionList.Set()
-		ol, err := newOptionList(options, strings.Join(options, ","))
-		if err != nil {
-			t.Errorf("test %d: unexpected error preparing test: %v", i, err)
-		}
-
-		if err := ol.Set(tt.opts); (err != nil) != tt.err {
-			t.Errorf("test %d: could not parse options as expected: %v", i, err)
-		}
-		if tt.ex != "" && tt.ex != ol.String() {
-			t.Errorf("test %d: resulting options not as expected: %s != %s",
-				i, tt.ex, ol.String())
-		}
-	}
-}
-
-var bfMap = map[string]int{
-	options[0]: 0,
-	options[1]: 1,
-	options[2]: 1 << 1,
-}
-
-func TestBitFlags(t *testing.T) {
-	tests := []struct {
-		opts     string
-		ex       int
-		parseErr bool
-		logicErr bool
-	}{
-		{
-			opts: "one,two",
-			ex:   3,
-		},
-		{ // Duplicate test
-			opts:     "zero,two,two",
-			ex:       -1,
-			parseErr: true,
-		},
-		{ // Not included test
-			opts:     "zero,two,three",
-			ex:       -1,
-			parseErr: true,
-		},
-		{ // Test 10 in 11
-			opts: "one,two",
-			ex:   1,
-		},
-		{ // Test 11 not in 01
-			opts:     "one",
-			ex:       3,
-			logicErr: true,
-		},
-	}
-
-	for i, tt := range tests {
-		// test NewBitFlags
-		if _, err := newBitFlags(options, tt.opts, bfMap); (err != nil) != tt.parseErr {
-			t.Errorf("test %d: unexpected error in newBitFlag: %v", i, err)
-		}
-
-		bf, err := newBitFlags(options, strings.Join(options, ","), bfMap)
-		if err != nil {
-			t.Errorf("test %d: unexpected error preparing test: %v", i, err)
-		}
-
-		// test BitFlags.Set()
-		if err := bf.Set(tt.opts); (err != nil) != tt.parseErr {
-			t.Errorf("test %d: Could not parse options as expected: %v", i, err)
-		}
-		if tt.ex >= 0 && bf.hasFlag(tt.ex) == tt.logicErr {
-			t.Errorf("test %d: Result was unexpected: %d != %d",
-				i, tt.ex, bf.flags)
-		}
-	}
-}
