+++ rkt/rkt.go
@@ -28,6 +28,7 @@ import (
 	"github.com/coreos/rkt/pkg/keystore"
 	"github.com/coreos/rkt/pkg/multicall"
 	"github.com/coreos/rkt/rkt/config"
+	rktflag "github.com/coreos/rkt/rkt/flag"
 )
 
 const (
@@ -35,27 +36,6 @@ const (
 	cliDescription = "rkt, the application container runner"
 
 	defaultDataDir = "/var/lib/rkt"
-)
-
-const (
-	insecureNone  = 0
-	insecureImage = 1 << (iota - 1)
-	insecureTls
-	insecureOnDisk
-
-	insecureAll = (insecureImage | insecureTls | insecureOnDisk)
-)
-
-var (
-	insecureOptions = []string{"none", "image", "tls", "ondisk", "all"}
-
-	insecureOptionsMap = map[string]int{
-		insecureOptions[0]: insecureNone,
-		insecureOptions[1]: insecureImage,
-		insecureOptions[2]: insecureTls,
-		insecureOptions[3]: insecureOnDisk,
-		insecureOptions[4]: insecureAll,
-	}
 )
 
 // flagInsecureSkipVerify is a deprecated flag that is equivalent to setting
@@ -71,9 +51,9 @@ func (sv *skipVerify) Set(s string) error {
 
 	switch v {
 	case false:
-		err = (*bitFlags)(globalFlags.InsecureFlags).Set("none")
+		err = (*rktflag.BitFlags)(globalFlags.InsecureFlags).Set("none")
 	case true:
-		err = (*bitFlags)(globalFlags.InsecureFlags).Set("all")
+		err = (*rktflag.BitFlags)(globalFlags.InsecureFlags).Set("all")
 	}
 	return err
 }
@@ -111,28 +91,6 @@ func (d *absDir) Type() string {
 	return "absolute-directory"
 }
 
-type secFlags bitFlags
-
-func (sf *secFlags) SkipImageCheck() bool {
-	return (*bitFlags)(sf).hasFlag(insecureImage)
-}
-
-func (sf *secFlags) SkipTlsCheck() bool {
-	return (*bitFlags)(sf).hasFlag(insecureTls)
-}
-
-func (sf *secFlags) SkipOnDiskCheck() bool {
-	return (*bitFlags)(sf).hasFlag(insecureOnDisk)
-}
-
-func (sf *secFlags) SkipAllSecurityChecks() bool {
-	return (*bitFlags)(sf).hasFlag(insecureAll)
-}
-
-func (sf *secFlags) SkipAnySecurityChecks() bool {
-	return (*bitFlags)(sf).flags != 0
-}
-
 var (
 	tabOut      *tabwriter.Writer
 	globalFlags = struct {
@@ -141,7 +99,7 @@ var (
 		LocalConfigDir     string
 		Debug              bool
 		Help               bool
-		InsecureFlags      *secFlags
+		InsecureFlags      *rktflag.SecFlags
 		TrustKeysFromHttps bool
 	}{
 		Dir:             defaultDataDir,
@@ -158,20 +116,20 @@ var cmdRkt = &cobra.Command{
 }
 
 func init() {
-	bf, err := newBitFlags(insecureOptions, "none", insecureOptionsMap)
+	bf, err := rktflag.NewBitFlags(rktflag.InsecureOptions, "none", rktflag.InsecureOptionsMap)
 	if err != nil {
 		stderr("rkt: problem initializing: %v", err)
 		os.Exit(1)
 	}
 
-	globalFlags.InsecureFlags = (*secFlags)(bf)
+	globalFlags.InsecureFlags = (*rktflag.SecFlags)(bf)
 	skipVerify := new(skipVerify)
 
 	cmdRkt.PersistentFlags().BoolVar(&globalFlags.Debug, "debug", false, "print out more debug information to stderr")
 	cmdRkt.PersistentFlags().Var((*absDir)(&globalFlags.Dir), "dir", "rkt data directory")
 	cmdRkt.PersistentFlags().Var((*absDir)(&globalFlags.SystemConfigDir), "system-config", "system configuration directory")
 	cmdRkt.PersistentFlags().Var((*absDir)(&globalFlags.LocalConfigDir), "local-config", "local configuration directory")
-	cmdRkt.PersistentFlags().Var((*bitFlags)(globalFlags.InsecureFlags), "insecure-options",
+	cmdRkt.PersistentFlags().Var((*rktflag.BitFlags)(globalFlags.InsecureFlags), "insecure-options",
 		fmt.Sprintf("comma-separated list of security features to disable. Allowed values: %s",
 			globalFlags.InsecureFlags.PermissibleString()))
 	cmdRkt.PersistentFlags().BoolVar(&globalFlags.TrustKeysFromHttps, "trust-keys-from-https",
